- Start checking distances from nebulae for Electricae Radialem` - Add a new rule for `PrimaryStar` - Stop requiring `Electricae Pluma` to be hear a nebula, start making them need the primary star to be one of `A,B,D,H,N,W` - Fix `atmosComp` criteria to use single word terms, eg: `SulphurDioxide` not `Sulphur dioxide` - Fix colour mapping for `Bacterium Nebulus` - `Stratum Araneamus` now can only predict colour variant `Emerald`

+++ SrvSurvey/net/spansh.cs
@@ -131,8 +131,8 @@ namespace SrvSurvey.net
         public long? age;
         public double? axialTilt;
         public int bodyId;
-        public long? distanceToArrival;
-        public long? semiMajorAxis;
+        public double? distanceToArrival;
+        public double? semiMajorAxis;
         public long? id64;
         public string luminosity;
         public bool? mainStar;
